add saving of Ethereum prices to database

+++ src/main/java/com/test/crypto/model/BinanceResponse.java
@@ -3,12 +3,14 @@ package com.test.crypto.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class BinanceResponse {
     private String symbol;
-    private Double bidPrice;
+    private BigDecimal bidPrice;
     private Double bidQty;
-    private Double askPrice;
+    private BigDecimal askPrice;
     private Double askQty;
 }
